move decode method to contract
